Add support for Ceylon Archives (.car files)
Ceylon Archives, Ceylon[1]'s binaries for the JVM backend, are mostly
Java Archives, with some added bits. `strip-nondeterminism -t jar`
already removes most of the nondeterminism in them, but some remains:

- META-INF/MANIFEST.MF contains a Bundle-Version, which is timestamped:

      Bundle-Version: 1.0.0.v20150101-1215

- META-INF/mapping.txt contains a comment with timestamp:

      #
      #Fri Nov 27 17:42:58 CET 2015

- META-INF/maven/.../pom.properties also contains a comment with
  timestamp:

      #Generated by Ceylon
      #Fri Nov 27 17:42:58 CET 2015

The new car handler removes those three sources of nondeterminism by
patching the Bundle-Version in the manifest file and stripping comments
from the other two files; otherwise, it delegates to the jar and zip
handlers.

If no timestamp is specified (-T option), the Bundle-Version is given a
default timestamp of 20100101-0000; work on the Java backend for Ceylon
started in July 2010, so there should be no real Ceylon Archives older
than that. The zip handler will use its own default date.

Note that strip-nondeterminism will not update the .sha1 file of the
.car file; you will need to do this yourself, for example with this
command:

    echo -n $(sha1sum X.car | cut -f1 -d' ') > X.car.sha1

1: http://ceylon-lang.org/